third_party/WebKit/LayoutTests/fast/dom/DeviceOrientation/window-property-expected.txt

Tests that the window.DeviceOrientationEvent and window.ondeviceorientation properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL typeof window.DeviceOrientationEvent == 'object' should be true. Was false.
FAIL typeof window.DeviceOrientationEvent == 'function' should be false. Was true.
PASS window.propertyIsEnumerable('DeviceOrientationEvent') is false
PASS 'DeviceOrientationEvent' in window is true
PASS window.hasOwnProperty('DeviceOrientationEvent') is true
PASS typeof window.ondeviceorientation == 'object' is true
PASS hasOnDeviceOrientationProperty() is true
PASS 'ondeviceorientation' in window is true
PASS window.hasOwnProperty('ondeviceorientation') is true
PASS successfullyParsed is true

TEST COMPLETE
